+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/BRuntime.java
@@ -20,25 +20,14 @@ import org.ballerinalang.jvm.observability.ObservabilityConstants;
 import org.ballerinalang.jvm.observability.ObserveUtils;
 import org.ballerinalang.jvm.observability.ObserverContext;
 import org.ballerinalang.jvm.scheduling.Scheduler;
-import org.ballerinalang.jvm.scheduling.State;
 import org.ballerinalang.jvm.scheduling.Strand;
 import org.ballerinalang.jvm.scheduling.StrandMetadata;
-import org.ballerinalang.jvm.types.BFunctionType;
 import org.ballerinalang.jvm.types.BTypes;
-import org.ballerinalang.jvm.values.ErrorValue;
-import org.ballerinalang.jvm.values.FPValue;
-import org.ballerinalang.jvm.values.FutureValue;
 import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.jvm.values.connector.CallableUnitCallback;
 
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.function.Supplier;
 
 /**
  * External API to be used by the interop users to control Ballerina runtime behavior.
@@ -63,111 +52,23 @@ public class BRuntime {
         return new BRuntime(strand.scheduler);
     }
 
-    /**
-     * Block the current strand to execute asynchronously.
-     *
-     * @return Future object to unblock the strand.
-     */
-    public static CompletableFuture<Object> markAsync() {
-        Strand strand = Scheduler.getStrand();
-        strand.blockedOnExtern = true;
-        strand.setState(State.BLOCK_AND_YIELD);
-        CompletableFuture<Object> future = new CompletableFuture<>();
-        future.whenComplete(new Unblocker(strand));
-        return future;
-    }
-
-    /**
-     * Invoke Function Pointer asynchronously. This will schedule the function and block the strand.
-     *
-     * @param func                 Function Pointer to be invoked.
-     * @param strandName           Name for newly creating strand which is used to execute the function pointer.
-     *                             This is optional and can be null.
-     * @param metadata             Meta data of new strand.
-     * @param args                 Ballerina function arguments.
-     * @param resultHandleFunction Function used to process the result received after execution of function.
-     * @return Future Value
-     */
-    public FutureValue invokeFunctionPointerAsync(FPValue<?, ?> func, String strandName, StrandMetadata metadata,
-                                                  Object[] args, Function<Object, Object> resultHandleFunction) {
-        AsyncFunctionCallback callback = new AsyncFunctionCallback() {
-            @Override
-            public void notifySuccess() {
-                setReturnValues(resultHandleFunction.apply(getFutureResult()));
-            }
-
-            @Override
-            public void notifyFailure(ErrorValue error) {
-                handleRuntimeErrors(error);
-            }
-        };
-        return invokeFunctionPointerAsync(func, strandName, metadata, args, callback);
-    }
-
-    /**
-     * Invoke Function Pointer asynchronously. This will schedule the function and block the strand.
-     *
-     * @param func       Function Pointer to be invoked.
-     * @param strandName Name for newly creating strand which is used to execute the function pointer. This is
-     *                   optional and can be null.
-     * @param metadata   Meta data of new strand.
-     * @param args       Ballerina function arguments.
-     * @param callback   Asynchronous call back.
-     * @return Future value
-     */
-    public FutureValue invokeFunctionPointerAsync(FPValue<?, ?> func, String strandName, StrandMetadata metadata,
-                                                  Object[] args, AsyncFunctionCallback callback) {
-
-        Strand strand = Scheduler.getStrand();
-        return invokeFunctionPointerAsync(func, strand, strandName, metadata, args, callback);
-    }
-
-    /**
-     * Invoke Function Pointer asynchronously given number of times. This will schedule the function and block the
-     * strand. This method can be used with collection of data where we need to invoke the function pointer for each
-     * item of the collection.
-     *
-     * @param func                 Function Pointer to be invoked.
-     * @param strandName           Name for newly creating strand which is used to execute the function pointer. This is
-     *                             optional and can be null.
-     * @param metadata             Meta data of new strand.
-     * @param noOfIterations       Number of iterations need to call the function pointer.
-     * @param argsSupplier         Supplier provides dynamic arguments to function pointer execution in each iteration.
-     * @param futureResultConsumer Consumer used to process the future value received after execution of function.
-     *                             Future value result will have the return object of the function pointer.
-     * @param returnValueSupplier  Suppler used to set the final return value for the parent function invocation.
-     */
-    public void invokeFunctionPointerAsyncIteratively(FPValue<?, ?> func, String strandName,
-                                                      StrandMetadata metadata, int noOfIterations,
-                                                      Supplier<Object[]> argsSupplier,
-                                                      Consumer<Object> futureResultConsumer,
-                                                      Supplier<Object> returnValueSupplier) {
-        if (noOfIterations <= 0) {
-            return;
-        }
-        Strand strand = Scheduler.getStrand();
-        blockStrand(strand);
-        AtomicInteger callCount = new AtomicInteger(0);
-        scheduleNextFunction(func, strand, strandName, metadata, noOfIterations, callCount, argsSupplier,
-                             futureResultConsumer, returnValueSupplier);
-    }
-
-    /**
-     * Invoke Object method asynchronously. This will schedule the function and block the strand.
-     *
-     * @param object     Object Value.
-     * @param methodName Name of the method.
-     * @param strandName Name for newly creating strand which is used to execute the function pointer. This is
-     *                   optional and can be null.
-     * @param metadata   Meta data of new strand.
-     * @param callback   Callback which will get notify once method execution done.
-     * @param args       Ballerina function arguments.
-     */
-    public void invokeMethodAsync(ObjectValue object, String methodName, String strandName, StrandMetadata metadata,
-                                  CallableUnitCallback callback, Object... args) {
-        Function<?, ?> func = o -> object.call((Strand) (((Object[]) o)[0]), methodName, args);
-        scheduler.schedule(new Object[1], func, null, callback, strandName, metadata);
-    }
+     /**
+      * Invoke Object method asynchronously. This will schedule the function and block the strand.
+      *
+      * @param object     Object Value.
+      * @param methodName Name of the method.
+      * @param strandName Name for newly creating strand which is used to execute the function pointer. This is optional
+      *                   and can be null.
+      * @param metadata   Meta data of new strand.
+      * @param callback   Callback which will get notify once method execution done.
+      * @param args       Ballerina function arguments.
+      * @return the result of the function invocation
+      */
+     public Object invokeMethodAsync(ObjectValue object, String methodName, String strandName, StrandMetadata metadata,
+                                     CallableUnitCallback callback, Object... args) {
+         Function<?, ?> func = o -> object.call((Strand) (((Object[]) o)[0]), methodName, args);
+         return scheduler.schedule(new Object[1], func, null, callback, strandName, metadata).result;
+     }
 
     /**
      * Invoke Object method asynchronously. This will schedule the function and block the strand.
@@ -193,148 +94,5 @@ public class BRuntime {
             return object.call(strand, methodName, args);
         };
         scheduler.schedule(new Object[1], func, null, callback, properties, BTypes.typeNull, strandName, metadata);
-    }
-
-    /**
-     * Invoke Object method synchronously. This will block the thread.
-     *
-     * @param object     Object Value.
-     * @param methodName Name of the method.
-     * @param strandName Name for newly creating strand which is used to execute the function pointer. This is
-     *                   optional and can be null.
-     * @param metadata   Meta data of new strand.
-     * @param args       Ballerina function arguments.
-     */
-    public void invokeMethodSync(ObjectValue object, String methodName, String strandName,
-                                 StrandMetadata metadata, Object... args) {
-        Function<?, ?> func = o -> object.call((Strand) (((Object[]) o)[0]), methodName, args);
-        Semaphore semaphore = new Semaphore(0);
-        final ErrorValue[] errorValue = new ErrorValue[1];
-        scheduler.schedule(new Object[1], func, null, new CallableUnitCallback() {
-            @Override
-            public void notifySuccess() {
-                semaphore.release();
-            }
-
-            @Override
-            public void notifyFailure(ErrorValue error) {
-                errorValue[0] = error;
-                semaphore.release();
-            }
-        }, strandName, metadata);
-        try {
-            semaphore.acquire();
-        } catch (InterruptedException e) {
-            // Ignore
-        }
-        if (errorValue[0] != null) {
-            throw errorValue[0];
-        }
-    }
-
-    /**
-     * Invoke Ballerina function and get the result.
-     *
-     * @param object     Ballerina object in which the function is defined.
-     * @param methodName Ballerina function name, to invoke.
-     * @param strandName Name for newly creating strand which is used to execute the function pointer. This is
-     *                   optional and can be null.
-     * @param metadata   Meta data of new strand.
-     * @param timeout    Timeout in milliseconds to wait until acquiring the semaphore.
-     * @param args       Ballerina function arguments.
-     * @return Ballerina function invoke result.
-     */
-    public Object getSyncMethodInvokeResult(ObjectValue object, String methodName, String strandName,
-                                            StrandMetadata metadata, int timeout, Object... args) {
-        Function<?, ?> func = o -> object.call((Strand) (((Object[]) o)[0]), methodName, args);
-        Semaphore semaphore = new Semaphore(0);
-        final ErrorValue[] errorValue = new ErrorValue[1];
-        // Add 1 more element to keep null for add the strand later.
-        Object[] params = new Object[]{null, args};
-        FutureValue futureValue = scheduler.schedule(params, func, null, new CallableUnitCallback() {
-            @Override
-            public void notifySuccess() {
-                semaphore.release();
-            }
-
-            @Override
-            public void notifyFailure(ErrorValue error) {
-                errorValue[0] = error;
-                semaphore.release();
-            }
-        }, strandName, metadata);
-        try {
-            semaphore.tryAcquire(timeout, TimeUnit.MILLISECONDS);
-        } catch (InterruptedException e) {
-            // Ignore
-        }
-        if (errorValue[0] != null) {
-            throw errorValue[0];
-        }
-        return futureValue.result;
-    }
-
-
-    private void scheduleNextFunction(FPValue<?, ?> func, Strand strand, String strandName,
-                                      StrandMetadata metadata, int noOfIterations,
-                                      AtomicInteger callCount, Supplier<Object[]> argsSupplier,
-                                      Consumer<Object> futureResultConsumer,
-                                      Supplier<Object> returnValueSupplier) {
-        AsyncFunctionCallback callback = new AsyncFunctionCallback() {
-            @Override
-            public void notifySuccess() {
-                futureResultConsumer.accept(getFutureResult());
-                if (callCount.incrementAndGet() != noOfIterations) {
-                    scheduleNextFunction(func, strand, strandName, metadata, noOfIterations, callCount, argsSupplier,
-                                         futureResultConsumer, returnValueSupplier);
-                } else {
-                    setReturnValues(returnValueSupplier.get());
-                }
-            }
-
-            @Override
-            public void notifyFailure(ErrorValue error) {
-                handleRuntimeErrors(error);
-            }
-        };
-        invokeFunctionPointerAsync(func, strand, strandName, metadata, argsSupplier.get(), callback);
-    }
-
-    private FutureValue invokeFunctionPointerAsync(FPValue<?, ?> func, Strand parent, String name,
-                                                   StrandMetadata metadata, Object[] args,
-                                                   AsyncFunctionCallback callback) {
-
-        blockStrand(parent);
-        final FutureValue future = scheduler.createFuture(parent, null, null,
-                                                          ((BFunctionType) func.getType()).retType, name, metadata);
-        future.callback = callback;
-        callback.setFuture(future);
-        callback.setStrand(parent);
-        return scheduler.scheduleLocal(args, func, parent, future);
-    }
-
-    private void blockStrand(Strand strand) {
-        if (!strand.blockedOnExtern) {
-            strand.blockedOnExtern = true;
-            strand.setState(State.BLOCK_AND_YIELD);
-            strand.returnValue = null;
-        }
-    }
-
-    private static class Unblocker implements java.util.function.BiConsumer<Object, Throwable> {
-
-        private Strand strand;
-
-        public Unblocker(Strand strand) {
-            this.strand = strand;
-        }
-
-        @Override
-        public void accept(Object returnValue, Throwable throwable) {
-            if (throwable == null) {
-                this.strand.returnValue = returnValue;
-                this.strand.scheduler.unblockStrand(strand);
-            }
-        }
     }
 }
