+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/values/FPValue.java
@@ -17,7 +17,8 @@
  */
 package org.ballerinalang.jvm.values;
 
-import org.ballerinalang.jvm.BRuntime;
+import org.ballerinalang.jvm.runtime.AsyncUtils;
+import org.ballerinalang.jvm.scheduling.Scheduler;
 import org.ballerinalang.jvm.scheduling.StrandMetadata;
 import org.ballerinalang.jvm.types.BType;
 import org.ballerinalang.jvm.util.BLangConstants;
@@ -66,8 +67,8 @@ public class FPValue<T, R> implements BFunctionPointer<T, R>, RefValue {
 
     public FutureValue asyncCall(Object[] args, Function<Object, Object> resultHandleFunction,
                                  StrandMetadata metaData) {
-        return BRuntime.getCurrentRuntime().invokeFunctionPointerAsync(this, this.strandName, metaData,
-                                                                       args, resultHandleFunction);
+        return AsyncUtils.invokeFunctionPointerAsync(this, this.strandName, metaData,
+                                                     args, resultHandleFunction, Scheduler.getStrand().scheduler);
     }
 
     public Function<T, R> getFunction() {
