+++ langlib/lang.table/src/main/java/org/ballerinalang/langlib/table/Filter.java
@@ -18,7 +18,8 @@
 
 package org.ballerinalang.langlib.table;
 
-import org.ballerinalang.jvm.BRuntime;
+import org.ballerinalang.jvm.runtime.AsyncUtils;
+import org.ballerinalang.jvm.scheduling.Scheduler;
 import org.ballerinalang.jvm.scheduling.Strand;
 import org.ballerinalang.jvm.scheduling.StrandMetadata;
 import org.ballerinalang.jvm.types.BTableType;
@@ -58,9 +59,9 @@ public class Filter {
         int size = tbl.size();
         AtomicInteger index = new AtomicInteger(-1);
 
-        BRuntime.getCurrentRuntime()
+        AsyncUtils
                 .invokeFunctionPointerAsyncIteratively(func, null, METADATA, size,
-                        () -> new Object[]{strand,
+                                                       () -> new Object[]{strand,
                                 tbl.get(tbl.getKeys()[index.incrementAndGet()]), true},
                         result -> {
                             if ((Boolean) result) {
@@ -68,7 +69,7 @@ public class Filter {
                                 Object value = tbl.get(key);
                                 newTable.put(key, value);
                             }
-                        }, () -> newTable);
+                        }, () -> newTable, Scheduler.getStrand().scheduler);
         return newTable;
     }
