+++ tests/jballerina-unit-test/src/test/java/org/ballerinalang/nativeimpl/jvm/tests/AsyncInterop.java
@@ -16,7 +16,7 @@
  */
 package org.ballerinalang.nativeimpl.jvm.tests;
 
-import org.ballerinalang.jvm.BRuntime;
+import org.ballerinalang.jvm.runtime.AsyncUtils;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -30,7 +30,7 @@ import java.util.concurrent.CompletableFuture;
 public class AsyncInterop {
 
     public static int countSlowly() {
-        CompletableFuture<Object> future = BRuntime.markAsync();
+        CompletableFuture<Object> future = AsyncUtils.markAsync();
 
         new Thread(() -> {
             sleep();
